Move UI Folder into TenshiTranslator
