+++ TenshiTranslator/UI/BatchSizeSlider.py
