+++ TenshiTranslator/UI/TimeoutSlider.py
